Minor repo-updater code cleanup (#56482)
Just some minor code cleanup from reading through the init code of repo-updater. Also finally removed some more global assignments from the enterprise split.
