+++ internal/repos/BUILD.bazel
@@ -87,6 +87,7 @@ go_library(
         "//internal/httptestutil",
         "//internal/jsonc",
         "//internal/lazyregexp",
+        "//internal/licensing",
         "//internal/limiter",
         "//internal/metrics",
         "//internal/observation",
@@ -190,6 +191,7 @@ go_test(
         "//internal/goroutine",
         "//internal/httpcli",
         "//internal/httptestutil",
+        "//internal/licensing",
         "//internal/limiter",
         "//internal/observation",
         "//internal/ratelimit",
