+++ internal/repos/syncer_test.go
@@ -19,6 +19,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/api"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/database"
+	"github.com/sourcegraph/sourcegraph/internal/database/dbmocks"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/awscodecommit"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/bitbucketcloud"
@@ -27,6 +28,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/gitlab"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/gitolite"
 	"github.com/sourcegraph/sourcegraph/internal/goroutine"
+	"github.com/sourcegraph/sourcegraph/internal/licensing"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
 	"github.com/sourcegraph/sourcegraph/internal/ratelimit"
 	"github.com/sourcegraph/sourcegraph/internal/repos"
@@ -2045,4 +2047,155 @@ func setupSyncErroredTest(ctx context.Context, s repos.Store, t *testing.T,
 
 var noopProgressRecorder = func(ctx context.Context, progress repos.SyncProgress, final bool) error {
 	return nil
+}
+
+func TestCreateRepoLicenseHook(t *testing.T) {
+	ctx := context.Background()
+
+	// Set up mock repo count
+	mockRepoStore := dbmocks.NewMockRepoStore()
+	mockStore := repos.NewMockStore()
+	mockStore.RepoStoreFunc.SetDefaultReturn(mockRepoStore)
+
+	tests := map[string]struct {
+		maxPrivateRepos int
+		unrestricted    bool
+		numPrivateRepos int
+		newRepo         *types.Repo
+		wantErr         bool
+	}{
+		"private repo, unrestricted": {
+			unrestricted:    true,
+			numPrivateRepos: 99999999,
+			newRepo:         &types.Repo{Private: true},
+			wantErr:         false,
+		},
+		"private repo, max private repos reached": {
+			maxPrivateRepos: 1,
+			numPrivateRepos: 1,
+			newRepo:         &types.Repo{Private: true},
+			wantErr:         true,
+		},
+		"public repo, max private repos reached": {
+			maxPrivateRepos: 1,
+			numPrivateRepos: 1,
+			newRepo:         &types.Repo{Private: false},
+			wantErr:         false,
+		},
+		"private repo, max private repos not reached": {
+			maxPrivateRepos: 2,
+			numPrivateRepos: 1,
+			newRepo:         &types.Repo{Private: true},
+			wantErr:         false,
+		},
+	}
+
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			mockRepoStore.CountFunc.SetDefaultReturn(test.numPrivateRepos, nil)
+
+			defaultMock := licensing.MockCheckFeature
+			licensing.MockCheckFeature = func(feature licensing.Feature) error {
+				if prFeature, ok := feature.(*licensing.FeaturePrivateRepositories); ok {
+					prFeature.MaxNumPrivateRepos = test.maxPrivateRepos
+					prFeature.Unrestricted = test.unrestricted
+				}
+
+				return nil
+			}
+			defer func() {
+				licensing.MockCheckFeature = defaultMock
+			}()
+
+			err := repos.CreateRepoLicenseHook(ctx, mockStore, test.newRepo)
+			if gotErr := err != nil; gotErr != test.wantErr {
+				t.Fatalf("got err: %t, want err: %t, err: %q", gotErr, test.wantErr, err)
+			}
+		})
+	}
+}
+
+func TestUpdateRepoLicenseHook(t *testing.T) {
+	ctx := context.Background()
+
+	// Set up mock repo count
+	mockRepoStore := dbmocks.NewMockRepoStore()
+	mockStore := repos.NewMockStore()
+	mockStore.RepoStoreFunc.SetDefaultReturn(mockRepoStore)
+
+	tests := map[string]struct {
+		maxPrivateRepos int
+		unrestricted    bool
+		numPrivateRepos int
+		existingRepo    *types.Repo
+		newRepo         *types.Repo
+		wantErr         bool
+	}{
+		"from public to private, unrestricted": {
+			unrestricted:    true,
+			numPrivateRepos: 99999999,
+			existingRepo:    &types.Repo{Private: false},
+			newRepo:         &types.Repo{Private: true},
+			wantErr:         false,
+		},
+		"from public to private, max private repos reached": {
+			maxPrivateRepos: 1,
+			numPrivateRepos: 1,
+			existingRepo:    &types.Repo{Private: false},
+			newRepo:         &types.Repo{Private: true},
+			wantErr:         true,
+		},
+		"from private to private, max private repos reached": {
+			maxPrivateRepos: 1,
+			numPrivateRepos: 1,
+			existingRepo:    &types.Repo{Private: true},
+			newRepo:         &types.Repo{Private: true},
+			wantErr:         false,
+		},
+		"from private to public, max private repos reached": {
+			maxPrivateRepos: 1,
+			numPrivateRepos: 1,
+			existingRepo:    &types.Repo{Private: true},
+			newRepo:         &types.Repo{Private: false},
+			wantErr:         false,
+		},
+		"from private deleted to private not deleted, max private repos reached": {
+			maxPrivateRepos: 1,
+			numPrivateRepos: 1,
+			existingRepo:    &types.Repo{Private: true, DeletedAt: time.Now()},
+			newRepo:         &types.Repo{Private: true, DeletedAt: time.Time{}},
+			wantErr:         true,
+		},
+		"from public to private, max private repos not reached": {
+			maxPrivateRepos: 2,
+			numPrivateRepos: 1,
+			existingRepo:    &types.Repo{Private: false},
+			newRepo:         &types.Repo{Private: true},
+			wantErr:         false,
+		},
+	}
+
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			mockRepoStore.CountFunc.SetDefaultReturn(test.numPrivateRepos, nil)
+
+			defaultMock := licensing.MockCheckFeature
+			licensing.MockCheckFeature = func(feature licensing.Feature) error {
+				if prFeature, ok := feature.(*licensing.FeaturePrivateRepositories); ok {
+					prFeature.MaxNumPrivateRepos = test.maxPrivateRepos
+					prFeature.Unrestricted = test.unrestricted
+				}
+
+				return nil
+			}
+			defer func() {
+				licensing.MockCheckFeature = defaultMock
+			}()
+
+			err := repos.UpdateRepoLicenseHook(ctx, mockStore, test.existingRepo, test.newRepo)
+			if gotErr := err != nil; gotErr != test.wantErr {
+				t.Fatalf("got err: %t, want err: %t, err: %q", gotErr, test.wantErr, err)
+			}
+		})
+	}
 }
